chore: Move testhelper to src
