+++ packages/openscd/src/test-helper/mock-edits.ts
