+++ packages/openscd/src/test-helper/mock-plugins.ts
